dom/src/threads/test/test_regExpStatics.html

<!DOCTYPE HTML>
<html>
<!--
Tests of DOM Worker Threads RegExp statics
-->
<head>
  <title>Test for DOM Worker Threads RegExp statics</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script class="testbody" type="text/javascript">

  const WORKER_COUNT = 25;

  function workerScript() {
    var runCount = 0;
    var timeout;

    this.messageListener = function(message, source) {
      run();
      timeout = setTimeout(run, 0);
      timeout = setTimeout(run, 5000);
    };

    function run() {
      if (RegExp.$1) {
        throw "RegExp.$1 already set!";
        cancelTimeout(timeout);
      }

      var match = /a(sd)f/("asdf");
      if (!RegExp.$1) {
        throw "RegExp.$1 didn't get set!";
        cancelTimeout(timeout);
      }

      if (++runCount == 3) {
        postMessageToPool("done");
      }
    }
  }

  var pool = navigator.newWorkerPool();

  var doneWorkers = 0;
  pool.messageListener = function(message, source) {
    if (++doneWorkers == WORKER_COUNT) {
      SimpleTest.finish();
    }
  };

  pool.errorListener = function(error, source) {
    ok(false, "Worker had an error: " + error);
    SimpleTest.finish();
  };

  for (var i = 0; i < WORKER_COUNT; i++) {
    var worker = pool.createWorker("(" + workerScript + ")();");
    worker.postMessage("start");
  }

  SimpleTest.waitForExplicitFinish();

</script>
</pre>
</body>
</html>
